Rename test helpers, update react jest config
Rename client/src/layouts/helpers to:
=> client/src/layouts/test-helpers
- ensures that it's clear that there's test fixtures/ utils inside

Rename ReduxWrap.tsx and router.helper.tsx to:
- ReduxWrap.helper.spec.tsx
- router.helper.spec.tsx
- update imports as needed

Update react.jest.config.js to ignore tests with/ in:
- *.helper.spec.tsx extension
- **/test-helpers/** directory

Remove the find -name -delete command in "heroku-prebuild" script due to updates to jest config
