+++ client/src/layouts/specs/Nav/TodoAppBar.spec.tsx
@@ -2,12 +2,12 @@ import React from 'react';
 
 import '@testing-library/jest-dom/extend-expect';
 import { cleanup } from '@testing-library/react';
-import renderWithRouter from '../../helpers/router.helper';
+import renderWithRouter from '../../test-helpers/router.helper.spec';
 
 import configureStore from '../../../store/configureStore';
 
 import TodoAppBar from '../../Nav/TodoAppBar';
-import ReduxWrap from '../../helpers/ReduxWrap';
+import ReduxWrap from '../../test-helpers/ReduxWrap.helper.spec';
 
 
 const store = configureStore();
