+++ client/src/layouts/test-helpers/router.helper.spec.tsx
@@ -45,4 +45,5 @@ const renderWithRouter = (ui: ReactElement, config: renderRouterFixtureConfig = 
   }
 }
 
+test('', () => expect(true).toBe(true))
 export default renderWithRouter;
