+++ client/src/pages/specs/RegisterPage.spec.tsx
@@ -5,7 +5,7 @@ import { HashRouter, Route, Redirect } from 'react-router-dom';
 // ----- Test Helpers
 import MockAdapter from 'axios-mock-adapter';
 import '@testing-library/jest-dom/extend-expect';
-import renderWithRouter from '../../layouts/helpers/router.helper';
+import renderWithRouter from '../../layouts/test-helpers/router.helper.spec';
 import { render, cleanup, waitForElement } from '@testing-library/react';
 
 import axios from '../../axios'
@@ -13,7 +13,7 @@ import configureStore from '../../store/configureStore'
 import RegisterPage from '../RegisterPage'
 
 import { StoreShape } from '../../types';
-import ReduxWrap from '../../layouts/helpers/ReduxWrap';
+import ReduxWrap from '../../layouts/test-helpers/ReduxWrap.helper.spec';
